removed unnecessary code
