--- app/src/main/res/layout/activity_search.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright (c) 2017. Matsuda, Akihit (akihito104)
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-    >
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        >
-        <android.support.design.widget.AppBarLayout
-            android:id="@+id/search_appbar_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            >
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/search_toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?android:attr/actionBarSize"
-                app:titleTextColor="@color/text_primary_inverse"
-                app:layout_collapseMode="pin"
-                />
-        </android.support.design.widget.AppBarLayout>
-
-        <FrameLayout
-            android:id="@+id/search_timeline_container"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:layout_behavior="@string/appbar_scrolling_view_behavior"
-            />
-
-        <com.freshdigitable.udonroad.ffab.IndicatableFFAB
-            android:id="@+id/ffab"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="@dimen/iffab_margin_bottom"
-            android:layout_gravity="center_horizontal|bottom"
-            android:visibility="invisible"
-            app:menu="@menu/iffab_main"
-            tools:visibility="visible"
-            />
-    </FrameLayout>
-</layout>
